Refactor logger response processing and update doc
